make player into component so it plays on all pages
