+++ package.json
@@ -23,7 +23,7 @@
     "intersection-observer": "0.12.2",
     "line-awesome": "1.3.0",
     "moment-timezone": "~0.5.45",
-    "next": "14.1.0",
+    "next": "^14.2.4",
     "pg": "8.11.3",
     "prop-types": "15.8.1",
     "react": "18.2.0",
